backend/chart_service/requirements.txt
fastapi
uvicorn
altair
pandas
python-multipart
vl-convert-python